# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API - Cadastro de Usuários",
    "description": "Documentação das rotas de cadastro e consulta de instituições beneficiárias e pessoas físicas.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/registerdonor": {
      "post": {
        "summary": "Cadastro de Doador",
        "description": "Cadastra um novo doador.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "namedonor": { "type": "string" },
                  "emaildonor": { "type": "string", "format": "email" },
                  "passworddonor": { "type": "string", "format": "password" }
                },
                "required": ["namedonor", "emaildonor", "passworddonor"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Cadastro realizado!" },
          "400": { "description": "Preencha todos os campos!" },
          "409": { "description": "❌ E-mail já cadastrado." },
          "500": { "description": "Erro ao cadastrar doador." }
        }
      }
    },
...
    "/pedidos": {
      "get": {
        "summary": "Listar todos os pedidos",
        "description": "Retorna todos os pedidos cadastrados no sistema.",
        "responses": {
          "200": {
            "description": "Lista completa de pedidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "pessoa_beneficiaria_id": { "type": "integer" },
                      "instituicao_id": { "type": "integer" },
                      "name_item": { "type": "string" },
                      "description": { "type": "string" },
                      "quantity_item": { "type": "integer" },
                      "category": { "type": "string" },
                      "urgencia_enum": { "type": "string" },
                      "locate": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      }
    },
    "/doador": {
      "get": {
        "summary": "Buscar doador por ID",
        "description": "Retorna os dados de um doador específico com base no ID. Requer autenticação com token JWT.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "ID do doador",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Doador encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "nome": { "type": "string" },
                    "email": { "type": "string" },
                    "telefone": { "type": "string" },
                    "cpf": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token de autenticação não fornecido"
          },
          "403": {
            "description": "Token inválido ou expirado"
          },
          "404": {
            "description": "Doador não encontrado"
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      }
    },
    "/deletedoador": {
    "delete": {
    "summary": "Deletar doador",
    "description": "Remove um doador do banco de dados com base no ID. Requer autenticação com token JWT.",
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "parameters": [
      {
        "name": "id",
        "in": "query",
        "required": true,
        "description": "ID do doador a ser deletado",
        "schema": {
          "type": "string"
        }
      }
    ],
    "responses": {
      "200": {
        "description": "Doador deletado com sucesso",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "mensagem": {
                  "type": "string",
                  "example": "Doador deletado com sucesso"
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "ID do doador não fornecido"
      },
      "404": {
        "description": "Doador não encontrado"
      },
      "500": {
        "description": "Erro no servidor"
      }
    }
  }
},
...
"/doadorup/{id}": {
  "put": {
    "summary": "Atualizar dados do doador",
    "description": "Atualiza o nome e email do doador. A atualização exige que a senha atual seja informada corretamente. Requer autenticação JWT.",
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID do doador a ser atualizado",
        "schema": {
          "type": "string"
        }
      }
    ],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "required": ["nameDoador", "emailDoador", "passwordDoador"],
            "properties": {
              "nameDoador": {
                "type": "string",
                "description": "Nome do doador"
              },
              "emailDoador": {
                "type": "string",
                "format": "email",
                "description": "Email do doador"
              },
              "passwordDoador": {
                "type": "string",
                "description": "Senha atual do doador para validação"
              }
            }
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Doador atualizado com sucesso",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "mensagem": {
                  "type": "string",
                  "example": "Doador atualizado com sucesso"
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "Campos obrigatórios não preenchidos"
      },
      "401": {
        "description": "Senha atual incorreta"
      },
      "404": {
        "description": "Doador não encontrado"
      },
      "500": {
        "description": "Erro no servidor"
      }
    }
  }
},
...
}
}